Scout external message handler
